Comments
This is a very basic stunt,  I thought of it on a train back
from an event last week.
I saved my GitHub page using a browser and then made a few
modifications to make it look more realistic as a Microsoft
project. I then hijacked the A href clicks and used them
to show the wonderfully awesome clippyjs by smore.

https://www.smore.com/clippy-js

I then uploaded the result to the gh-pages branch of a Git Repo
as I felt this would trick the average user into thinking that
the code was a genuine GitHub Project.